requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.6.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.5.0
pyyaml>=6.0
aiofiles>=23.0
websockets>=12.0
python-multipart>=0.0.22
httpx>=0.25.0
rich>=13.0
python-dotenv>=1.0.0
